test(benchmark): address review feedback
