+++ .github/workflows/tests.yml
@@ -32,34 +32,31 @@ jobs:
 
         name: Feed Generation Benchmarks
 
-        env:
-            BENCHMARK_IMAGE: composer:2.10.2@sha256:5946476338742b200bb9ff88f8be56275ddae4b3949c72305cb0dbf10cfcb760
-
         steps:
             -   name: Checkout code
                 uses: actions/checkout@3d3c42e5aac5ba805825da76410c181273ba90b1
                 with:
                     persist-credentials: false
 
-            -   name: Verify benchmark runtime
+            -   name: Setup PHP
+                uses: shivammathur/setup-php@f3e473d116dcccaddc5834248c87452386958240
+                with:
+                    php-version: 8.5
+                    extensions: bcmath, curl, dom, iconv, libxml, mbstring, pdo, pdo_sqlite, zip
+                    coverage: none
+                    tools: composer:2.10.2
+
+            -   name: Install dependencies
                 run: |
-                    test "$(docker run --rm "$BENCHMARK_IMAGE" php -r 'echo PHP_VERSION;')" = "8.5.8"
-                    test "$(docker run --rm "$BENCHMARK_IMAGE" composer --version --no-ansi | awk '{print $3}')" = "2.10.2"
-                    test "$(docker run --rm "$BENCHMARK_IMAGE" php -r 'echo extension_loaded("xdebug") ? "yes" : "no";')" = "no"
+                    composer install --no-interaction --no-progress --prefer-dist --no-scripts
+                    composer clear
+                    composer prepare
 
             -   name: Verify regression snapshots
                 run: test "$(git ls-files -- '.benchmarks/**/*.snap' | wc -l)" -eq 5
 
             -   name: Execute feed generation benchmarks
-                run: |
-                    docker run --rm \
-                        --cpus=1 \
-                        --memory=1g \
-                        --memory-swap=1g \
-                        --volume "$GITHUB_WORKSPACE:/app" \
-                        --workdir /app \
-                        "$BENCHMARK_IMAGE" \
-                        sh -lc 'composer install --no-interaction --no-progress --prefer-dist --no-scripts --ignore-platform-req=ext-bcmath && composer clear && composer prepare && composer test:benchmark'
+                run: composer test:benchmark
 
             -   name: Ensure regression snapshots are unchanged
                 run: |
